+++ odoo/local-src/connector_esb/tests/test_export_product.py
@@ -2,7 +2,7 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import tools, fields
+from odoo import tools
 from .common import ESBXMLTestCase
 
 
@@ -215,16 +215,7 @@ class ExportProductTestCase(ESBXMLTestCase):
             self.assertDictEqual(mapper.map_record(rec).values(), expected)
 
     def test_filename(self):
-        today = fields.Date.today().replace('-', '')
-        with self.backend.work_on(self.model._name,
-                                  timestamp=self.timestamp) as work:
-            expected = 'Product_{}.xml'.format(today)
-            writer = work.component(usage='local.xml.writer')
-            self.assertEqual(
-                writer.filename(), expected)
-            writer = work.component(usage='sftp.xml.writer')
-            self.assertEqual(
-                writer.filename(), expected)
+        self.check_filename('Product_{}.xml')
 
     @tools.mute_logger('dicttoxml')
     def test_record_exporter_local(self):